disallows actual method calls in ghost code
