+++ src/main/scala/viper/gobra/frontend/info/implementation/typing/MemberTyping.scala
@@ -10,6 +10,7 @@ import org.bitbucket.inkytonik.kiama.util.Messaging.{Messages, error, noMessages
 import viper.gobra.GoVerifier
 import viper.gobra.ast.frontend._
 import viper.gobra.frontend.Config
+import viper.gobra.frontend.info.base.SymbolTable.MPredicateSpec
 import viper.gobra.frontend.info.implementation.TypeInfoImpl
 import viper.gobra.util.Constants
 
@@ -55,6 +56,37 @@ trait MemberTyping extends BaseTyping { this: TypeInfoImpl =>
       }
     case s: PActualStatement =>
       wellDefStmt(s).out
+
+    case ip: PImplementationProof =>
+      val subType = symbType(ip.subT)
+      val superType = symbType(ip.superT)
+
+      val syntaxImplementsMsgs = syntaxImplements(subType, superType).asReason(ip, s"${ip.subT} does not implement the interface ${ip.superT}")
+      if (syntaxImplementsMsgs.nonEmpty) syntaxImplementsMsgs
+      else {
+        addDemandedImplements(subType, superType)
+
+        {
+          val badReceiverTypes = ip.memberProofs.map(m => miscType(m.receiver))
+            .filter(t => !identicalTypes(t, subType))
+          error(ip, s"The receiver of all methods included in the implementation proof must be $subType, " +
+            s"but encountered: ${badReceiverTypes.distinct.mkString(", ")}", cond = badReceiverTypes.nonEmpty)
+        } ++ {
+          val superPredNames = memberSet(superType).collect { case (n, m: MPredicateSpec) => (n, m) }
+          val allPredicatesDefined = PropertyResult.bigAnd(superPredNames.map { case (name, symb) =>
+            val valid = tryMethodLikeLookup(subType, PIdnUse(name)).isDefined ||
+              ip.alias.exists(al => al.left.name == name)
+            failedProp({
+              val argTypes = symb.args map symb.context.typ
+
+              s"predicate $name is not defined for type $subType. " +
+                s"Either declare a predicate 'pred ($subType) $name(${argTypes.mkString(", ")})' " +
+                s"or declare a predicate 'pred p($subType${if (argTypes.isEmpty) "" else ", "}${argTypes.mkString(", ")})' with some name p and add 'pred $name := p' to the implementation proof."
+            }, !valid)
+          })
+          allPredicatesDefined.asReason(ip, "Some predicate definitions are missing")
+        }
+      }
   }
 
   private def wellDefConstSpec(spec: PConstSpec): Messages = {
