+++ src/main/scala/viper/gobra/frontend/info/implementation/typing/ghost/GhostMemberTyping.scala
@@ -26,37 +26,6 @@ trait GhostMemberTyping extends BaseTyping { this: TypeInfoImpl =>
       body.fold(noMessages)(assignableToSpec) ++
         isReceiverType.errors(miscType(receiver))(member) ++
         nonVariadicArguments(args)
-
-    case ip: PImplementationProof =>
-      val subType = symbType(ip.subT)
-      val superType = symbType(ip.superT)
-
-      val syntaxImplementsMsgs = syntaxImplements(subType, superType).asReason(ip, s"${ip.subT} does not implement the interface ${ip.superT}")
-      if (syntaxImplementsMsgs.nonEmpty) syntaxImplementsMsgs
-      else {
-        addDemandedImplements(subType, superType)
-
-        {
-          val badReceiverTypes = ip.memberProofs.map(m => miscType(m.receiver))
-            .filter(t => !identicalTypes(t, subType))
-          error(ip, s"The receiver of all methods included in the implementation proof must be $subType, " +
-            s"but encountered: ${badReceiverTypes.distinct.mkString(", ")}", cond = badReceiverTypes.nonEmpty)
-        } ++ {
-          val superPredNames = memberSet(superType).collect{ case (n, m: MPredicateSpec) => (n, m) }
-          val allPredicatesDefined = PropertyResult.bigAnd(superPredNames.map{ case (name, symb) =>
-            val valid = tryMethodLikeLookup(subType, PIdnUse(name)).isDefined ||
-              ip.alias.exists(al => al.left.name == name)
-            failedProp({
-              val argTypes = symb.args map symb.context.typ
-
-              s"predicate $name is not defined for type $subType. " +
-                s"Either declare a predicate 'pred ($subType) $name(${argTypes.mkString(", ")})' " +
-                s"or declare a predicate 'pred p($subType${if (argTypes.isEmpty) "" else ", "}${argTypes.mkString(", ")})' with some name p and add 'pred $name := p' to the implementation proof."
-            }, !valid)
-          })
-          allPredicatesDefined.asReason(ip, "Some predicate definitions are missing")
-        }
-      }
   }
 
   private[typing] def wellFoundedIfNeeded(member: PMember): Messages = {
@@ -111,7 +80,7 @@ trait GhostMemberTyping extends BaseTyping { this: TypeInfoImpl =>
     }
   }
 
-  private[ghost] def isPureBlock(block: PBlock): Messages = {
+  private[typing] def isPureBlock(block: PBlock): Messages = {
     block.nonEmptyStmts match {
       case Vector(PReturn(Vector(ret))) => isPureExpr(ret)
       case b => error(block, s"For now, the body of a pure block is expected to be a single return with a pure expression, got $b instead")
